+++ src/main/php/ConsoleInputStream.php
@@ -5,9 +5,9 @@
  * For the full copyright and license information, please view the LICENSE
  * file that was distributed with this source code.
  *
- * @package  net\stubbles\console
+ * @package  stubbles\console
  */
-namespace net\stubbles\console;
+namespace stubbles\console;
 use stubbles\streams\DecodingInputStream;
 use stubbles\streams\ResourceInputStream;
 /**