+++ src/test/php/net/stubbles/console/ConsoleOutputStreamTestCase.php
@@ -5,11 +5,11 @@
  * For the full copyright and license information, please view the LICENSE
  * file that was distributed with this source code.
  *
- * @package  net\stubbles\console
+ * @package  stubbles\console
  */
-namespace net\stubbles\console;
+namespace stubbles\console;
 /**
- * Test for net\stubbles\console\ConsoleOutputStream.
+ * Test for stubbles\console\ConsoleOutputStream.
  *
  * @group  console
  */