+++ src/test/php/net/stubbles/console/input/BrokeredUserInput.php
@@ -5,7 +5,7 @@
  * For the full copyright and license information, please view the LICENSE
  * file that was distributed with this source code.
  *
- * @package  net\stubbles\console
+ * @package  stubbles\console
  */
 namespace org\stubbles\console\test;
 /**