+++ CHANGELOG.md
@@ -1,3 +1,15 @@
+3.0.0 (2014-06-??)
+------------------
+
+### BC breaks
+
+  * removed namespace prefix `net`, base namespace is now `stubbles\console` only
+
+### Other changes
+
+  * upgraded to stubbles/core 4.x
+
+
 2.6.0 (2013-10-28)
 ------------------
 
@@ -8,7 +20,7 @@
 2.5.0 (2013-10-26)
 ------------------
 
-  * remove script for compiling console apps to phars, use https://packagist.org/packages/clue/phar-composer instead
+  * removed script for compiling console apps to phars, use https://packagist.org/packages/clue/phar-composer instead
 
 
 2.4.0 (2013-10-25)